# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: AI Forex Trading Bot – REST API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/trades:
    get:
      operationId: getTrades
      tags: [trades]
      summary: Paginated list of trades
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, cancelled, all]
            default: all
        - name: pair
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Paginated trades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTrades"

    post:
      operationId: createTrade
      tags: [trades]
      summary: Open a new manual trade record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Created trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /v1/trades/stats:
    get:
      operationId: getTradeStats
      tags: [trades]
      summary: Aggregate trade statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Trade statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeStats"
  /v1/trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a single trade by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found

    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a trade (SL/TP/notes)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"

    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Cancel / delete a trade record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Signals ───────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Trade:
      type: object
      required: [id, pair, direction, entryPrice, stopLoss, takeProfit, lotSize, status, createdAt]
      properties:
        id:
          type: string
        pair:
          type: string
        direction:
          type: string
          enum: [buy, sell]
        entryPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        lotSize:
          type: number
        status:
          type: string
          enum: [open, closed, cancelled]
        pnl:
          type: ["number", "null"]
        riskRewardRatio:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        signalId:
          type: ["string", "null"]
        openedAt:
          type: ["string", "null"]
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    TradeInput:
      type: object
      required: [pair, direction, entryPrice, stopLoss, takeProfit, lotSize]
      properties:
        pair:
          type: string
        direction:
          type: string
          enum: [buy, sell]
        entryPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        lotSize:
          type: number
        notes:
          type: string
        signalId:
          type: string
    TradeUpdate:
      type: object
      properties:
        stopLoss:
          type: number
        takeProfit:
          type: number
        lotSize:
          type: number
        notes:
          type: string
        status:
          type: string
          enum: [closed, cancelled]
        pnl:
          type: number
        closedAt:
          type: string
          format: date-time
    TradeStats:
      type: object
      required: [totalTrades, winningTrades, losingTrades, winRate, totalPnl, avgWin, avgLoss, profitFactor, maxDrawdown, avgRr]
      properties:
        totalTrades:
          type: number
        winningTrades:
          type: number
        losingTrades:
          type: number
        winRate:
          type: number
        totalPnl:
          type: number
        avgWin:
          type: number
        avgLoss:
          type: number
        profitFactor:
          type: number
        maxDrawdown:
          type: number
        avgRr:
          type: number
    PaginatedTrades:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: number
        page:
          type: number
        limit:
          type: number

    # Signals
...